Grey's Initial Commit
Example Commit Message:

Add a cube to the sample scene, add a Materials folder, and add a blue material. Add the blue material to the cube.

Commit messages should always have this tense, "add, delete, change" so on, so on. You are explaining what the other members of the team will get when they open your changes. Don't say "I added this and this, or "adds" or "will add". Your messages don't need to be as in-depth as my example, it just explains the structure: what you did and where you did it. The where is important for this project since we will likely be working in different places often and it helps to say where the changes can be found.